Large refactor: ASM now holds InsnList, List<TryCatchBlockNodes>
This is a patch-version bump since these fields (and the field it
replaces) were already internal.

This opens us up to injection of InsnLists so that instructions don't
have to be parsed in-order, which lets us have nicer try-catch blocks in
future. (Since we need to jump past the handlers before they've been
created)